+++ Projet4/v3/App/templates/manage_users.php
@@ -1,25 +1,40 @@
 <?php $this->title = "P4 v3 - page Gestion des utilisateurs"; ?>
 
 <section id="pageBlock">
-    <div class="manageUsersBlock">
-        <p>
-            <?= $this->session->show('delete_user'); ?>
-        </p>
-        <h2>Gérer les comptes utilisateurs</h2>
-        <ul>
-            <?php
-            foreach ($users as $user)
-            {
-                if ($user->getRole() === 'user') {
-                ?>
-                    <li>
-                        <?= htmlspecialchars($user->getPseudo());?> - <?= htmlspecialchars($user->getRole());?> - crée le <?= htmlspecialchars($user->getCreatedAt());?> 
-                        <a href="../public/index.php?route=deleteUser&amp;userId=<?= $user->getId(); ?>">Supprimer</a>
-                    </li>
-                <?php
-                }
-            }
-            ?>
-        </ul>       
+    <p>
+        <?= $this->session->show('delete_user'); ?>
+    </p>
+    <div class="espacePersoBlock">
+        <?php include 'admin_nav.php'; ?>
+        <div class="espacePersoBlockInfo">
+            <div class="manageUsersBlock">
+            <h2>Gérer les comptes utilisateurs</h2>
+                <ul>
+                    <?php
+                    if (count($users) <= 1) {
+                        ?>
+                        <p>Aucun utilisateur inscrit</p>
+                        <?php
+                    } else {
+                        foreach ($users as $user)
+                        {
+                            if ($user->getRole() === 'user') {
+                            ?>
+                                <li>
+                                    <div class="manageUserContent">
+                                        <?= htmlspecialchars($user->getPseudo());?> - <?= htmlspecialchars($user->getRole());?> - crée le<span class="comDate"><?= htmlspecialchars($user->getCreatedAt());?></span>
+                                    </div>
+                                    <div class="manageComLinks">
+                                        <a href="../public/index.php?route=deleteUser&amp;userId=<?= $user->getId(); ?>">Supprimer</a>
+                                    </div>
+                                </li>
+                            <?php
+                            }
+                        }
+                    }
+                    ?>
+                </ul>
+            </div>
+        </div>
     </div>
 </section>
